int is enough.

+++ x-pack/plugin/autoscaling/src/main/java/org/elasticsearch/xpack/autoscaling/shards/FrozenShardsDeciderService.java
@@ -53,15 +53,15 @@ public class FrozenShardsDeciderService implements AutoscalingDeciderService {
     @Override
     public AutoscalingDeciderResult scale(Settings configuration, AutoscalingDeciderContext context) {
         // we assume that nodes do not grow beyond 64GB here.
-        long shards = countFrozenShards(context.state().metadata());
+        int shards = countFrozenShards(context.state().metadata());
         long memory = shards * MEMORY_PER_SHARD.get(configuration).getBytes();
         return new AutoscalingDeciderResult(AutoscalingCapacity.builder().total(null, memory).build(), new FrozenShardsReason(shards));
     }
 
-    static long countFrozenShards(Metadata metadata) {
+    static int countFrozenShards(Metadata metadata) {
         return StreamSupport.stream(metadata.spliterator(), false)
             .filter(imd -> isFrozenIndex(imd.getSettings()))
-            .mapToLong(IndexMetadata::getTotalNumberOfShards)
+            .mapToInt(IndexMetadata::getTotalNumberOfShards)
             .sum();
     }
